Backend/requirements.txt
# Flask Framework
Flask==3.0.3
Flask-CORS==4.0.0

# Database
pymongo==4.6.1

# Firebase
firebase-admin==6.4.0

# Environment variables
python-dotenv==1.0.1

# HTTP requests
requests==2.32.3

# Utilities
python-dateutil==2.8.2
pytz==2023.4